Ah... Finally.

+++ FallenRobot/modules/sql/global_bans_sql.py
@@ -1,12 +1,13 @@
 import threading
 
+from sqlalchemy import Boolean, Column, BigInteger, String, UnicodeText
+
 from FallenRobot.modules.sql import BASE, SESSION
-from sqlalchemy import Boolean, Column, Integer, String, UnicodeText
 
 
 class GloballyBannedUsers(BASE):
     __tablename__ = "gbans"
-    user_id = Column(Integer, primary_key=True)
+    user_id = Column(BigInteger, primary_key=True)
     name = Column(UnicodeText, nullable=False)
     reason = Column(UnicodeText)
